chore(monorepo): remove dependency on connect in components and device-utils
