+++ packages/components/package.json
@@ -26,9 +26,9 @@
         "@suite-common/validators": "workspace:*",
         "@testing-library/jest-dom": "^6.6.3",
         "@trezor/asset-utils": "workspace:*",
-        "@trezor/connect": "workspace:*",
         "@trezor/dom-utils": "workspace:*",
         "@trezor/env-utils": "workspace:*",
+        "@trezor/protobuf": "workspace:*",
         "@trezor/react-utils": "workspace:*",
         "@trezor/styles": "workspace:*",
         "@trezor/theme": "workspace:*",
