+++ packages/components/src/components/animations/DeviceAnimation.tsx
@@ -4,7 +4,7 @@ import styled, { useTheme } from 'styled-components';
 
 import { DEFAULT_FLAGSHIP_MODEL } from '@suite-common/suite-constants';
 import { getNarrowedDeviceModelInternal } from '@suite-common/suite-utils';
-import { DeviceModelInternal } from '@trezor/connect';
+import { DeviceModelInternal } from '@trezor/protobuf';
 
 import { AnimationWrapper, Shape } from './AnimationPrimitives';
 import { resolveStaticPath } from '../../utils/resolveStaticPath';
