+++ packages/components/src/components/animations/LottieAnimation.tsx
@@ -5,7 +5,8 @@ import styled from 'styled-components';
 
 import { DEFAULT_FLAGSHIP_MODEL } from '@suite-common/suite-constants';
 import { getNarrowedDeviceModelInternal } from '@suite-common/suite-utils';
-import { DeviceModelInternal } from '@trezor/connect';
+// todo: components should not rely on @trezor/protobuf
+import { DeviceModelInternal } from '@trezor/protobuf';
 
 import { AnimationWrapper, Shape } from './AnimationPrimitives';
 import { resolveStaticPath } from '../../utils/resolveStaticPath';
