+++ packages/connect/src/types/firmware.ts
@@ -1,3 +1,4 @@
+import { VersionArray } from '@trezor/device-utils';
 import { DeviceModelInternal } from '@trezor/protobuf';
 import { Type } from '@trezor/schema-utils';
 
@@ -8,8 +9,6 @@ export type FirmwareRange = Record<
         max: string;
     }
 >;
-
-export type VersionArray = [number, number, number];
 
 export type FirmwareRelease = {
     required: boolean;
