+++ packages/connect/tsconfig.json
@@ -8,6 +8,7 @@
         { "path": "../connect-analytics" },
         { "path": "../connect-common" },
         { "path": "../crypto-utils" },
+        { "path": "../device-utils" },
         { "path": "../protobuf" },
         { "path": "../protocol" },
         { "path": "../schema-utils" },
