+++ packages/device-utils/src/index.ts
@@ -1,4 +1,4 @@
 export * from './firmwareUtils';
 export * from './bootloaderUtils';
-export * from './modelUtils';
 export * from './modeUtils';
+export * from './types';
