+++ packages/device-utils/src/types.ts
@@ -1,8 +1,80 @@
-import { Device } from '@trezor/connect';
-
-export type PartialDevice = {
-    features?: Device['features'];
-    firmwareType?: Device['firmwareType'];
-};
-
 export type FirmwareVersionString = `${number}.${number}.${number}`;
+
+export enum FirmwareType {
+    BitcoinOnly = 'bitcoin-only',
+    Regular = 'regular',
+}
+
+export type VersionArray = [number, number, number];
+
+export type FeaturesNarrowing =
+    | {
+          major_version: 2;
+          minor_version: number;
+          patch_version: number;
+          fw_major: null;
+          fw_minor: null;
+          fw_patch: null;
+          bootloader_mode: true;
+          firmware_present: false;
+      }
+    | {
+          major_version: 2;
+          minor_version: number;
+          patch_version: number;
+          fw_major: null;
+          fw_minor: null;
+          fw_patch: null;
+          bootloader_mode: null;
+          firmware_present: null;
+      }
+    | {
+          major_version: 2;
+          minor_version: number;
+          patch_version: number;
+          fw_major: 2;
+          fw_minor: number;
+          fw_patch: number;
+          bootloader_mode: true;
+          firmware_present: true;
+      }
+    | {
+          major_version: 1;
+          minor_version: number;
+          patch_version: number;
+          fw_major: null;
+          fw_minor: null;
+          fw_patch: null;
+          bootloader_mode: true;
+          firmware_present: false;
+      }
+    | {
+          major_version: 1;
+          minor_version: number;
+          patch_version: number;
+          fw_major: null;
+          fw_minor: null;
+          fw_patch: null;
+          bootloader_mode: true;
+          firmware_present: true;
+      };
+
+// todo: this is copy-pasted from packages/protobuf/src/messages
+export type PartialDevice = {
+    firmwareType?: FirmwareType;
+
+    features?: {
+        major_version: number;
+        minor_version: number;
+        patch_version: number;
+        bootloader_mode: boolean | null;
+        initialized: boolean | null;
+        revision: string | null;
+        bootloader_hash: string | null;
+        fw_major: number | null;
+        fw_minor: number | null;
+        fw_patch: number | null;
+        no_backup: boolean | null;
+        unit_btconly?: boolean;
+    };
+};
