+++ packages/device-utils/tsconfig.json
@@ -1,5 +1,5 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [{ "path": "../connect" }]
+    "references": []
 }
