+++ packages/product-components/src/components/ConfirmOnDevice/ConfirmOnDeviceContent.tsx
@@ -3,7 +3,7 @@ import { ReactNode } from 'react';
 import styled, { css } from 'styled-components';
 
 import { Column, IconButton, Row, Text } from '@trezor/components';
-import { DeviceModelInternal } from '@trezor/connect';
+import { DeviceModelInternal } from '@trezor/protobuf';
 import { borders, spacings, spacingsPx } from '@trezor/theme';
 
 import { RotateDeviceImage } from '../RotateDeviceImage/RotateDeviceImage';
