+++ packages/product-components/src/components/PassphraseTypeCard/PassphraseTypeCard.tsx
@@ -5,8 +5,8 @@ import styled, { css } from 'styled-components';
 
 import { formInputsMaxLength } from '@suite-common/validators';
 import { TooltipProps } from '@trezor/components';
-import { DeviceModelInternal, Features } from '@trezor/connect';
 import { setCaretPosition } from '@trezor/dom-utils';
+import { DeviceModelInternal } from '@trezor/protobuf';
 import { useKeyPress } from '@trezor/react-utils';
 import { borders, spacingsPx } from '@trezor/theme';
 import { countBytesInString } from '@trezor/utils';
@@ -64,7 +64,7 @@ export type PassphraseTypeCardProps = {
     offerPassphraseOnDevice?: boolean;
     singleColModal?: boolean;
     deviceModel?: DeviceModelInternal;
-    deviceBackup?: Features['backup_type'] | null;
+    deviceBackup?: string | null;
     onSubmit: (value: string, passphraseOnDevice?: boolean) => void;
     learnMoreTooltipOnClick?: TooltipProps['addon'];
     learnMoreTooltipAppendTo?: TooltipProps['appendTo'];
