+++ packages/product-components/tsconfig.json
@@ -22,10 +22,11 @@
         {
             "path": "../../suite-common/wallet-utils"
         },
+        { "path": "../blockchain-link-types" },
         { "path": "../components" },
-        { "path": "../connect" },
         { "path": "../dom-utils" },
         { "path": "../env-utils" },
+        { "path": "../protobuf" },
         { "path": "../react-utils" },
         { "path": "../theme" },
         { "path": "../urls" },
