+++ packages/suite/src/utils/device/modelUtils.ts
