+++ packages/suite/tsconfig.json
@@ -102,6 +102,7 @@
         { "path": "../env-utils" },
         { "path": "../ipc-proxy" },
         { "path": "../product-components" },
+        { "path": "../protobuf" },
         { "path": "../react-utils" },
         { "path": "../styles" },
         { "path": "../suite-analytics" },
